# Improve performance of `permutation_of_terms`, `monomials`