--- a/src/Modules/UngradedModules/FreeModElem.jl
+++ b/src/Modules/UngradedModules/FreeModElem.jl
@@ -24,7 +24,7 @@ end
 
 function FreeModElem(i::Int, x::T, parent::FreeMod{T}) where T
   @assert 1 <= i <= rank(parent)
-  sparse_coords = sparse_row(base_ring(parent), [i], [x]; sort=false)
+  sparse_coords = sparse_row(base_ring(parent), [i], [x])
   return FreeModElem{T}(sparse_coords, parent)
 end
 
@@ -157,7 +157,7 @@ function basis(F::AbstractFreeMod)
   bas = Vector{elem_type(F)}(undef, dim(F))
   e = one(base_ring(F))
   for i=1:dim(F)
-    s = sparse_row(base_ring(F), [i], [e]; sort=false)
+    s = sparse_row(base_ring(F), [i], [e])
     bas[i] = F(s)
   end
   return bas
@@ -181,7 +181,7 @@ Return the `i`th basis vector of `F`, that is, return the `i`th standard unit ve
 function basis(F::AbstractFreeMod, i::Int)
   @assert 0 < i <= ngens(F)
   e = one(base_ring(F))
-  s = sparse_row(base_ring(F), [i], [e]; sort=false)
+  s = sparse_row(base_ring(F), [i], [e])
   return F(s)
 end
 gen(F::AbstractFreeMod, i::Int) = basis(F,i)