--- a/src/Modules/UngradedModules/SubquoModuleElem.jl
+++ b/src/Modules/UngradedModules/SubquoModuleElem.jl
@@ -935,7 +935,7 @@ Return the generators of `M`.
 function gens(M::SubquoModule{T}) where T
   R = base_ring(M)
   e = R(1)
-  return [SubquoModuleElem{T}(sparse_row(R, [i], [e]; sort=false), M) for i in 1:ngens(M)]
+  return [SubquoModuleElem{T}(sparse_row(R, [i], [e]), M) for i in 1:ngens(M)]
 end
 
 @doc raw"""
@@ -945,7 +945,7 @@ Return the `i`th generator of `M`.
 """
 function gen(M::SubquoModule{T}, i::Int) where T
   R = base_ring(M)
-  v = sparse_row(R, [i], [R(1)]; sort=false)
+  v = sparse_row(R, [i], [R(1)])
   return SubquoModuleElem{T}(v, M)
 end