Resolves #3055 (CC @HechtiDerLachs @simonbrandhorst)

Speed and memory usage are improved drastically. The main contributing factor is that now `permutation_of_terms` methods can fully take advantage of the type of the monomial resp. module orderings. This then avoids a ton of dynamic dispatches during runtime.

In addition, two smaller changes contribute:
- using `Iterators.reverse` instead of `reverse` avoids allocating a new vector when we just want to iterate backwards over an existing one
- changing `struct FreeModElem` to fully specify the type of its member `coords` gets rid of some additional dynamic dispatch.

Consider this setup:

    R, (x, y, z) = QQ[:x, :y, :z]

    m = ideal(R, gens(R))
    pow_m = m^200;

    F = FreeMod(R, 5)

    M, _ = pow_m*F;

    w = sum(gens(M));
    v = ambient_representative(w);
    g = sum(gens(pow_m));

Then we measure the performance before and after this PR.

Before:

    julia> @btime collect(monomials(v));
      738.822 ms (10193202 allocations: 393.99 MiB)

    julia> @btime collect(monomials(g));
      64.825 ms (1031634 allocations: 37.52 MiB)

    julia> @btime permutation_of_terms(v, default_ordering(parent(v)));
      616.583 ms (7253642 allocations: 263.95 MiB)

    julia> @btime permutation_of_terms(g, default_ordering(parent(g)));
      62.254 ms (950427 allocations: 34.42 MiB)

After:

    julia> @btime collect(monomials(v));
      226.134 ms (2436256 allocations: 126.02 MiB)

    julia> @btime collect(monomials(g));
      20.471 ms (81211 allocations: 3.41 MiB)

    julia> @btime permutation_of_terms(v, default_ordering(parent(v)));
      153.673 ms (133 allocations: 5.21 MiB)

    julia> @btime permutation_of_terms(g, default_ordering(parent(g)));
      18.163 ms (4 allocations: 317.34 KiB)