@fingolfin 2024-05-03
I believe that further changes of this kind would allow us to *drastically* reduce the memory usage of the algebraic geometry code (and probably some of our other code, too). E.g. I already pointed out elsewhere that conversion between OSCAR and Singular polynomials incurs up to 2 orders of magnitudes more allocations than I would expect.

I wish Julia would give some more guide rails here to avoid these pitfalls. It took me some hours to finally understand what's going on (admittedly the fact I only worked at this late at night likely didn't help). 

I guess tools like JET might have helped, but we currently have soooo many things it reports (and quite some "false positives", too) that it seems hopeless to find the worst offenders this way (well, at least for me -- perhaps someone who has mastered it better can do better? We'd just need to find such a person).


BTW I also experimented with https://github.com/JuliaLang/AllocCheck.jl here which helped a bit (it found the `reverse` problem) but no tool pointed me at the main culprit: dynamic dispatch to the comparison functions -- and I *still* don't really understand why that caused so many allocations -- that's bad. I.e. I can't fully explain why this PR is so effective in reducing allocations. If someone else can fill that in (e.g. @benlorenz @lgoettgens @thofma ...), please.