@thofma 2024-05-03
I think the allocations with dynamic dispatch are due to boxing of the return value (even for immutables). Last time I checked out JET, it gave waaaay too many false positives, often pointing to some Base methods. I played around with the filtering, but I could not make it work to my satisfaction. This was maybe two years ago. Maybe the situation has improved?