@fingolfin 2024-05-03
Just converted it to a draft. I have another factor 2 improvements in allocations for `collect(monomials(v))` as well (from 2.4 million down to 1.2 million). Will push update later

Perhaps I'll just remove the SRow related bit and put that into another PR, its contribution is relatively minor.